Revert "Moving the Stream Event Record classes into a cdap-formats module"
