+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/format/GrokRecordFormat.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.cdap.format;
+package co.cask.cdap.data.format;
 
 import co.cask.cdap.api.common.Bytes;
 import co.cask.cdap.api.data.format.StructuredRecord;
