+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/format/StreamEventRecordFormat.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2015 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.format;
+package co.cask.cdap.data.format;
 
 import co.cask.cdap.api.data.format.RecordFormat;
 import co.cask.cdap.api.flow.flowlet.StreamEvent;
