+++ cdap-data-fabric/src/main/resources/cdap/grok/patterns/junos
