+++ cdap-data-fabric/src/test/java/co/cask/cdap/data/format/AvroRecordFormatTest.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.format;
+package co.cask.cdap.data.format;
 
 import co.cask.cdap.api.common.Bytes;
 import co.cask.cdap.api.data.format.FormatSpecification;
@@ -23,6 +23,7 @@ import co.cask.cdap.api.data.format.RecordFormat;
 import co.cask.cdap.api.data.format.StructuredRecord;
 import co.cask.cdap.api.data.schema.Schema;
 import co.cask.cdap.api.flow.flowlet.StreamEvent;
+import co.cask.cdap.common.conf.Constants;
 import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -223,8 +224,8 @@ public class AvroRecordFormatTest {
 
     Map<String, String> headers = Maps.newHashMap();
     if (writeSchema) {
-      headers.put(AvroRecordFormat.SCHEMA, schemaString);
-      headers.put(AvroRecordFormat.SCHEMA_HASH,
+      headers.put(Constants.Stream.Headers.SCHEMA, schemaString);
+      headers.put(Constants.Stream.Headers.SCHEMA_HASH,
                   Hashing.md5().hashString(schemaString, Charsets.UTF_8).toString());
     }
