+++ cdap-explore/src/main/java/co/cask/cdap/hive/stream/StreamSerDe.java
@@ -21,9 +21,9 @@ import co.cask.cdap.api.data.schema.Schema;
 import co.cask.cdap.api.data.schema.UnsupportedTypeException;
 import co.cask.cdap.api.flow.flowlet.StreamEvent;
 import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.data.format.RecordFormats;
+import co.cask.cdap.data.format.StreamEventRecordFormat;
 import co.cask.cdap.data2.transaction.stream.StreamConfig;
-import co.cask.cdap.format.RecordFormats;
-import co.cask.cdap.format.StreamEventRecordFormat;
 import co.cask.cdap.hive.context.ContextManager;
 import co.cask.cdap.hive.serde.ObjectDeserializer;
 import co.cask.cdap.proto.Id;
