+++ pom.xml
@@ -2041,7 +2041,6 @@
       <modules>
         <module>cdap-common-unit-test</module>
         <module>cdap-api</module>
-        <module>cdap-formats</module>
         <module>cdap-common</module>
         <module>cdap-watchdog-api</module>
         <module>cdap-kafka</module>
